refactor: CheckedState Running 패키지로 이동
